+++ CHANGELOG.md
@@ -4,9 +4,12 @@
 
 ### Added
 
+ * convert command for FASTA/FASTQ/CSV/TSV file conversion, in place of the
+   more limited tab2seq command ([#14])
  * Rhesus germline HV and HJ allele FASTAs from
    [10.4049/jimmunol.1800342](https://doi.org/10.4049/jimmunol.1800342) ([#13])
 
+[#14]: https://github.com/ShawHahnLab/igseq/pull/14
 [#13]: https://github.com/ShawHahnLab/igseq/pull/13
 
 ## 0.1.1 - 2021-12-07